Add all of my work
I treated this whole thing like a game jam (even though it wasn't),
which meant I was cutting corners left and right. I really want to go
back and clean up some of this code, but I may never do that. It's
better to check this in and at least have bad code online than to have
no code online.