[blocked] Add testing environment / Separate projects (#60)
* Enable code analysis

This adds .Net Analyzers, StyleCop and SonarAnalyzer. All of them are configurable through `.editorconfig` and will be populated over time once the codebase grows

* Disable Globalization rules

* Thread-safe singleton, apply code style rules

* Logic separation

* Apply code style rules

* Small refactor, accessibility change

* Apply code style rules

* Apply code style rules

* Apply code style rules

* Apply code style rules

Readonly Properties are declared with getter only

* Apply code style rules

* Apply code style rules

* Add rule: warn on unnecessary usings

* Apply code style rules

* **Partially** apply code style rules

Partially, because the rest has to be checked separately inside the scene - if it still works as intended

* Add utility class for null reference checks

Note: this is just for the exported fields, they can often be null due to the missing assignment of an object in the scene

* Apply code style rules - address null references from the scene fields

* Reorder/rename fields, assign tilemap in `_Ready()`

`_lobbyTilemap` had to be initialized in the constructor, but since there is no constructor and we can't access a property before initialization, we can safely do this as soon as it enters the scene. That way we ensure it will assign the `_lobbyTilemap` for other class members

* Add extra info

* Rewrite nullable json case

* Update code analysis rules

* Update code analysis rules

* Apply code style rules

This commit is big, but this is only the code analysis part - reordering and renaming members, line spacing and explicit typing.

* Apply code style rules and refactor

* Revert PlayerData casing, throw on json errors, refactor jumper a bit

* Update code analysis rules

* Ignore Visual Studio generated files

* Re-enable .NET analyzers

* Revert PlayerData casing, ignore analyzers until refactored

* Leave an important note

* Update notice, remove default value

* Update comment

* Adjust the result text positioning

* Switch from `Any()` to Count check

* Remove duplicate character change

* Remove remaining vars

* Add comment

* Ignore generated contents

* Add NUnit project

* Move old tests to a temporary directory

* Add a placeholder test

* Add global NUnit usings

* Change package versions

* Initial Testing README

* Separate projects

The game had to be placed outside of the Test project to not have duplicate assembly warnings and `using`s collisions.

* No need to mark classes as internal here

* Migrate `ChatCommandParser` tests

* Enable code analysis in tests

* Move solution to the root

* Collect available commands

* Add partial tests

* Update comment

* Migrate the rest of CommandMatcher tests

* Migrate old Jumo command tests, remove old tests

* Update XMLDocs

* Explain nullable

* Merge #58

* Document Testing

* Temporarily add a workflow for notifying about depending pulls

* Update action

* Remove workflow

This required a manual action, never actually paid attention to that when using it, so we don't need it here 🤔

* Remove workflow

This required a manual action in order to re-run the jobs (inside action or a commit), never actually paid attention to that when using it, so we don't need it here 🤔

* Add a test workflow for Build and Test

* Typo in the version...

* Add missing link to Testing README

* Update .NET version

I'm also checking in the save file since we played a test game.

---------

Co-authored-by: Adam Learns <60950402+AdamLearns@users.noreply.github.com>
Co-authored-by: Adam <adam@adamlearns.live>