Refactoring, part 3 - Jumpers (#85)
* Initial commit

* Disable documentation rule

* Extract command handling logic from the Arena

* Merge #82

These are the same changes as in issue #82

* Refactor character choice

Removes method argument, sets the character choice directly on the player data and the method call was moved from Join to jumper Init

* Rename methods

* Rename leftovers from testing

* Initial commit

* (test) push StyleCop rule violation

* Revert StyleCop rule violation

* Move Twitch Constants

* Ignore URL string rule

We use `res`, so we get an annoying warning about hardcoded URLs

* Add Json constants/helpers

* Migrate Json handler from Arena to separate PlayerStats

This makes an external Json loader, which accepts a path to the resource, which is a Json file with player stats. This helps testing the player loading/saving.

`AllPlayerData` was moved to `PlayerStats`

* Match `.editorconfig` rules

* Remove init accessor

It does not matter in this context anyway

* Update conditions/cleanup

* Expose player loading and call when requested

* Add initial json tests

This is only to test if it actually passes on CI probably due to environment differences (paths)

* Update comment

* Add random Int to the Rng

* Add more Json tests

* Update comments, add `OneTimeTearDown`

* Revert OneTimeTearDown

Path is visible in the console and CI output.

* Convert `PlayerStats` into singleton

* Update StyleCop rules

* (temp) Update StyleCop rules

* (StyleCop) disable bracket rules

This is specific to the formatter for readability purposes (as much as other rules)

* Fix spacing rule violation (documentation)

* Fix typo

* Add new exceptions

* Add PlayerData Storing method, throw more specific exception

* Update SetUp, add new tests

* Ignore Coverlet

The test results come from Coverlet package that contain code coverage results, we don't need that

* Add missing coverage test

* Don't access player data directly

* Add fake class for random PlayerData

* Remove unnecessary null check

* Add new helper exception

* Adjust to new methods, add exceptions

* (StyleCop) Reformat string to fix false-positive

* Initial `Jumper` extraction

* Formatting (?)...

* Move Height converter to Jumpers, add Highest Jumper lookup

* Update dummy names

* Update comment

* Extract stats computing

* Format

* Remove sanity check and `StorePlayer`

I don't know why I added this, you can just overwrite when accessing directly by the index, we don't need a duplicate record guard

* Change arguments for the dummy players

* Rename `Jumpers`

* Small comment updates

* Remove unused exception classes

* Fix bug with getting the highest jumper

The camera wouldn't scroll because we were finding the highest Y value
instead of the lowest one. When there's only one player, this would have
seemed to work.

* Fix bug where name color is null in JSON

When that's the case, the player can't be loaded in correctly. To work
around this, the name color is set to white upon initializing (if it
would have been null).

* Fix bug with typing commands without joining

In that case, your jumper would be null, and trying to access it would
throw an exception.

---------

Co-authored-by: Adam Learns <60950402+AdamLearns@users.noreply.github.com>
Co-authored-by: Adam Damiano <adam@adamlearns.live>