Reorder built-ins, cache components, set player data on _Ready
Rather than accessing the nodes all the time through GetNode, just cache them once on _Ready and reuse.

This also changes when the player-specific Setters are fired, because when we call Init on join, the node has not really entered the tree yet, so the components were not ready to use.