Automatically change drawing objects
based on the current drawing `Y` we will change the single block - the
rest will come later as the platforms are made of 3 different tiles, so
the atlas coords object needs a refactor along with the coords picker
(hardcoded switch).

`RemovePoint` method is for now unused until the layer switch gets added
(next commit). This commit also fixes the drawing at certain heights as
the `Y` goes negative upwards and the actual arena starts at positive
numbers, which was a bit confusing to figure out at first. The `0` is
above the screen, then the next levels come after every 65 tiles.