Refactor ArenaBuilder, add object types
This change will make the `ArenaBuilder` operate on the object types now
instead, the separate game objects are defined in separate records and
stored in dictionary, where the `ArenaBuilder` fetches the atlas coords
from. The atlas coords remain declared as before, except they are just
used as names for convenience.