Refactor + change Modulation
This change mostly extracts magic numbers and caches the components.

This change also introduces new colors for timer overlay, which only
differ in Alpha channel being slightly reduced. The reasoning for this
is that when players stay behind the timer, they are completely covered
and you can't see if anyone is standing behind it when they are at the
top.