Gfx update (#96)
* Remove unused method

This method had no references, I don't know what was the origin of this, probably some early Jumper color overlay(?)

* Retrieve the node once, extract and refactor animation switch

* Migrate sprite rotation from new codebase

* Extract and reorder methods

* Extract physics methods, change sprite flipping logic

* Extract physics methods, change sprite flipping logic

* Add new platforms to the atlas

This also configures the physics layer to include the new objects
(blocks) and platforms.

* Initial arena builder commit

I have to migrate the code between machines, so committing the initial
stage of the Arena Builder class, which will automatically choose which
tile to put onto the arena based on the Y-coordinate. This code will
later be changed to shorten the game object declaration.

* Automatically change drawing objects

based on the current drawing `Y` we will change the single block - the
rest will come later as the platforms are made of 3 different tiles, so
the atlas coords object needs a refactor along with the coords picker
(hardcoded switch).

`RemovePoint` method is for now unused until the layer switch gets added
(next commit). This commit also fixes the drawing at certain heights as
the `Y` goes negative upwards and the actual arena starts at positive
numbers, which was a bit confusing to figure out at first. The `0` is
above the screen, then the next levels come after every 65 tiles.

* Update remaining tiles (add `OneWay`)

* Refactor ArenaBuilder, add object types

This change will make the `ArenaBuilder` operate on the object types now
instead, the separate game objects are defined in separate records and
stored in dictionary, where the `ArenaBuilder` fetches the atlas coords
from. The atlas coords remain declared as before, except they are just
used as names for convenience.

* Change the remaining platform drawers

Replaces the old `AddPlatform` with platform drawer from `ArenaBuilder`

* Auto ID update

* Fix a rendering bug on older machines

As in the node description:
https://github.com/godotengine/godot/issues/58314

The actual sprite should not be the zeroth child, otherwise the timer
overlay background will not move.

* Make platforms even thicker

This will just make them stand out more.

* Simplify the drawable object names

this also changes Clay->Terracotta, this sounds like a more fitting
name.

* Extract the object type switching points

* Tone down the main sprites a bit

* Add new test sprites

If they won't look good on stream, we will most likely remove them

* Tweak the texture change heights

The last one was too far for regular average height achieved by players,
it'd be almost never shown.

* Fix incorrect small platform drawing

When the platform was too small, e.g. 1 width, the right side was being
drawn over the left.

* Add more randomization to end screen platforms

There are less platforms now, so you can actually jump around.
Previously you would almost never fall down to the bottom.

* Shorten the maximum starting point

Platforms are now 1 tile longer

* Document `GetObjectType`

* Merge the code to avoid conflict

* Merge UID to avoid conflicts

* Adjust for conflict fix

* Prevent revival before match ends (#25)

* Don't revive before certain amount of time

This will simply abort the execution when the game timer reaches certain
value, but won't automatically refund the channel points - this has to
be done through Twitch Client, but this will require some
work+broadcaster token, so this part will be omitted for now.

* Remove unnecessary field

* Remove redundant comment

Things can be reverted from history, no need to keep this as a note.

* Fix typo

* Move comment to summary

This gives more clarity to the `width` argument.

* Update broken uids

* Apply review suggestion

This removes the cryptic switch

* Fix Godot's warning about invalid uid

Godot made a new uid for the Jumper resource, but it was not referenced
in the scene file for some reason :thinking:

* Document `BaseObject`

This was needed, because most likely the `Point` argument introduced
some confusion as to what it represented.

* Remove unnecessary classes

This instantiates new bases directly in the `GameObject` class,
resulting in less clutter. The extra classes were not needed anyway, I
forgot why I did that in the first place, I was supposed to do something
with the specific object types at some point