Add instruction bots that automatically jump
I didn't feel like properly extracting this, because I don't think we
actually pay much attention to the code quality in this project. This
change spawns 3 fake players that automatically jump after 2 seconds,
and after another 2 seconds they are reset to the initial position set
in each *bot*.

Since we make them jump with a command, the game will fade their names
out, so we have to flash them. Even if the name fades out a little, I
don't think it's a problem anyway, the names are perfectly readable so
we can keep flashing. It's probably a better solution than changing the
jumper just to add a flag that tells if the name has to be flashed,
which will only be used by the bots.