UI update (#95)
* Remove unused method

This method had no references, I don't know what was the origin of this, probably some early Jumper color overlay(?)

* Retrieve the node once, extract and refactor animation switch

* Migrate sprite rotation from new codebase

* Extract and reorder methods

* Extract physics methods, change sprite flipping logic

* Extract physics methods, change sprite flipping logic

* Update UI elements

This adds a TileSet for drawing UI boxes as backgrounds for various
labels. This was easier than designing a whole new UI with fixed size
windows. This way the boxes can just be redrawn to match new content.

* Update font settings

This makes the scaled names and UI text more pretty, but still has to be
tested on stream

* Add extra image with examples

Not really necessary, but more visuals is always better

* Add game banner

* Fix letter color, sprinkle some pixels on border

`90` on right side had its color changed from the other image editor
:eyes:

* Update LobbyOverlay

- Changes the position in the arena
- Adds character selection tips
- Changes the info in the labels
- Fixes number offset in the examples
- Adds Twitch Subscriber cosmetics info
- Lowers the font size in lobby overlay to add more text

* Refactor + change Modulation

This change mostly extracts magic numbers and caches the components.

This change also introduces new colors for timer overlay, which only
differ in Alpha channel being slightly reduced. The reasoning for this
is that when players stay behind the timer, they are completely covered
and you can't see if anyone is standing behind it when they are at the
top.

* Extract component lookup

I know this is a micro-optimization, but there is no need to constantly
perform the lookup. It's more about the unnecessary pollution.

* Replace `char` choices table with image

The characters will most likely never change anyway, plus, the image is
easy to change later.

* Adjust for conflict fix

* Prevent revival before match ends (#25)

* Don't revive before certain amount of time

This will simply abort the execution when the game timer reaches certain
value, but won't automatically refund the channel points - this has to
be done through Twitch Client, but this will require some
work+broadcaster token, so this part will be omitted for now.

* Remove unnecessary field

* Gfx update (#26)

* Make the UI more simple

- Removed extra instructions on screen
- Removed extra images (not used anymore)
- Added a new resource for UI backgrounds (StyleBoxTexture override)
- Added preset backgrounds to the overlays
- Repositioned `LobbyOverlay` - always centered
- Repositioned `EndScreenOverlay` - always top-left with offset

* Update wording on lobby overlay

Arena automatically offsets the position.

* Add instruction bots that automatically jump

I didn't feel like properly extracting this, because I don't think we
actually pay much attention to the code quality in this project. This
change spawns 3 fake players that automatically jump after 2 seconds,
and after another 2 seconds they are reset to the initial position set
in each *bot*.

Since we make them jump with a command, the game will fade their names
out, so we have to flash them. Even if the name fades out a little, I
don't think it's a problem anyway, the names are perfectly readable so
we can keep flashing. It's probably a better solution than changing the
jumper just to add a flag that tells if the name has to be flashed,
which will only be used by the bots.

* Resolve conflict from `gfx-update`

* Resolve `Arena.cs` conflict from `gfx-update`

* Resolve resource conflicts from `gfx-update`

* Update UI elements

This adds a TileSet for drawing UI boxes as backgrounds for various
labels. This was easier than designing a whole new UI with fixed size
windows. This way the boxes can just be redrawn to match new content.

* Update font settings

This makes the scaled names and UI text more pretty, but still has to be
tested on stream

* Add extra image with examples

Not really necessary, but more visuals is always better

* Add game banner

* Fix letter color, sprinkle some pixels on border

`90` on right side had its color changed from the other image editor
:eyes:

* Update LobbyOverlay

- Changes the position in the arena
- Adds character selection tips
- Changes the info in the labels
- Fixes number offset in the examples
- Adds Twitch Subscriber cosmetics info
- Lowers the font size in lobby overlay to add more text

* Extract component lookup

I know this is a micro-optimization, but there is no need to constantly
perform the lookup. It's more about the unnecessary pollution.

* Replace `char` choices table with image

The characters will most likely never change anyway, plus, the image is
easy to change later.

* Gfx update (#26)

* Make the UI more simple

- Removed extra instructions on screen
- Removed extra images (not used anymore)
- Added a new resource for UI backgrounds (StyleBoxTexture override)
- Added preset backgrounds to the overlays
- Repositioned `LobbyOverlay` - always centered
- Repositioned `EndScreenOverlay` - always top-left with offset

* Update wording on lobby overlay

Arena automatically offsets the position.

* Add instruction bots that automatically jump

I didn't feel like properly extracting this, because I don't think we
actually pay much attention to the code quality in this project. This
change spawns 3 fake players that automatically jump after 2 seconds,
and after another 2 seconds they are reset to the initial position set
in each *bot*.

Since we make them jump with a command, the game will fade their names
out, so we have to flash them. Even if the name fades out a little, I
don't think it's a problem anyway, the names are perfectly readable so
we can keep flashing. It's probably a better solution than changing the
jumper just to add a flag that tells if the name has to be flashed,
which will only be used by the bots.

* Resolve conflict from `gfx-update`

* Resolve resource conflicts from `gfx-update`

* Apply formatting

* Hide the game overlay on start

I believe this was set to visible at some point in the scene file and
left in by mistake, because we already have the code in the arena that
makes the `GameOverlay` visible after the lobby counter :thinking:

* Prevent accessing deleted bots

* Make instructions clearer

I cut down on the verbosity and added colors.

---------

Co-authored-by: Adam Damiano <adam@adamlearns.live>